Extract shortcode updates to separate class Tools\Shortcodes_Filter
